feat: guestManage 로직을 event 하위로 이동
